% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_r_coloc.R
\name{calc_credible_set}
\alias{calc_credible_set}
\title{Title}
\usage{
calc_credible_set(results, dat_join_to, coverage, result_type = "finemap.abf")
}
\arguments{
\item{results}{.}

\item{dat_join_to}{.}

\item{coverage}{.}

\item{result_type}{.}
}
\value{
.
}
\description{
Title
}
