% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{disk_cache}
\alias{disk_cache}
\title{Create a disk cache}
\usage{
disk_cache()
}
\value{
a [cachem::cache_disk()] object
}
\description{
Returns a [cachem::cache_disk()] object with the cache file path located in this packages
`inst/cache` folder
}
