% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_ebi_eqtl_catalogue.R
\name{get_ebi_eqtl_info}
\alias{get_ebi_eqtl_info}
\title{EBI eQTL Catalogue dataset information}
\usage{
get_ebi_eqtl_info()
}
\value{
a data.frame of study information
}
\description{
EBI eQTL Catalogue dataset information
}
