% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_ebi_gwas_catalogue.R
\name{get_ebi_gwas_info}
\alias{get_ebi_gwas_info}
\title{EBI GWAS Catalogue dataset information}
\usage{
get_ebi_gwas_info()
}
\value{
a data.frame of study information
}
\description{
EBI GWAS Catalogue dataset information
}
