% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebi_eqtl_catalogue.R
\name{get_ebi_study_info}
\alias{get_ebi_study_info}
\title{QBI eQTL Catalogue dataset information}
\usage{
get_ebi_study_info()
}
\value{
a data.frame of study information
}
\description{
QBI eQTL Catalogue dataset information
}
