% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_ensembl.R
\name{get_ensembl_gene_info}
\alias{get_ensembl_gene_info}
\title{Ensembl gene information}
\usage{
get_ensembl_gene_info(gene_id, build)
}
\arguments{
\item{gene_id}{a string, gene ID e.g. ENSG00000112164}

\item{build}{a string, one of c("GRCh37", "GRCh38")}
}
\value{
a list of gene information - see https://grch37.rest.ensembl.org/lookup/id/ docs for structure
}
\description{
Query the ensembl REST API for gene information
}
