% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_r_coloc.R
\name{get_harmonised_ld}
\alias{get_harmonised_ld}
\title{Title}
\usage{
get_harmonised_ld(dat, ref_path, chr, from, to, method)
}
\arguments{
\item{dat}{.}

\item{ref_path}{.}

\item{chr}{.}

\item{from}{.}

\item{to}{.}

\item{method}{.}
}
\value{
.
}
\description{
Title
}
