% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_ebi_eqtl_catalogue.R
\name{import_ebi_eqtl}
\alias{import_ebi_eqtl}
\title{Import EBI eQTL data}
\usage{
import_ebi_eqtl(
  study_info_table,
  studies = NULL,
  tissues = NULL,
  method = NULL,
  chr = NULL,
  bp_start = NULL,
  bp_end = NULL,
  gene_id = NULL,
  pthresh = NULL,
  build = "GRCh37"
)
}
\arguments{
\item{study_info_table}{.}

\item{studies}{.}

\item{tissues}{.}

\item{method}{.}

\item{chr}{.}

\item{bp_start}{.}

\item{bp_end}{.}

\item{gene_id}{.}

\item{pthresh}{.}

\item{build}{.}
}
\value{
a data.frame
}
\description{
A function to download from the EBI eQTL API
Note that there is some unexpected behaviour with the version2 filtering.
I thought that filtering by position and gene would be an `&` operation,
but it seems to be an `|` combination as when used together you get gene
eQTLs from way outside your specified region. Therefore I do the gene
filtering after.
}
