% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_ebi_eqtl_catalogue.R
\name{import_ebi_eqtl_dataset}
\alias{import_ebi_eqtl_dataset}
\title{Import EBI eQTL dataset}
\usage{
import_ebi_eqtl_dataset(id, chr, bp_start, bp_end)
}
\arguments{
\item{id}{the EBI eQTL study ID}

\item{chr}{.}

\item{bp_start}{.}

\item{bp_end}{.}
}
\value{
a data.table, the eQTL dataset with standardised column names
}
\description{
Import EBI eQTL dataset
}
