% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_ebi_gwas_catalogue.R
\name{import_ebi_gwas}
\alias{import_ebi_gwas}
\title{Import EBI GWAS associations}
\usage{
import_ebi_gwas(
  study_id,
  chr = NULL,
  bp_start = NULL,
  bp_end = NULL,
  build = "GRCh38",
  pthresh = NULL
)
}
\arguments{
\item{study_id}{.}

\item{chr}{.}

\item{bp_start}{.}

\item{bp_end}{.}

\item{build}{.}

\item{pthresh}{.}
}
\value{
a data.frame
}
\description{
A function to download from the EBI GWAS Catalogue. Build GRCh38 dbSNP Build 154
}
