% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_ieu_opengwas.R
\name{import_ieu_gwas}
\alias{import_ieu_gwas}
\title{Import IEU GWAS associations}
\usage{
import_ieu_gwas(
  study_id,
  chr = NULL,
  bp_start = NULL,
  bp_end = NULL,
  build = "GRCh37",
  pthresh = NULL
)
}
\arguments{
\item{study_id}{.}

\item{chr}{.}

\item{bp_start}{.}

\item{bp_end}{.}

\item{build}{.}

\item{pthresh}{.}
}
\value{
a data.frame
}
\description{
A function to download from the IEU OpenGWAS collection.
}
