% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_r_coloc.R
\name{make_coloc_dataset}
\alias{make_coloc_dataset}
\title{Coloc package standard dataset}
\usage{
make_coloc_dataset(
  dat,
  type,
  sdY = NA_real_,
  N = NA_real_,
  S = NA_real_,
  ld = NULL
)
}
\arguments{
\item{dat}{a data.table, with at least columns c("RSID","EA","OA","BP","BETA","SE") +/- c("EAF","N"),}

\item{type}{a string, either c("quant","cc")}

\item{sdY}{a numeric, std of the variable}

\item{N}{an integer, the sample size}

\item{S}{a numeric, proportion of cases}

\item{ld}{an LD matrix}
}
\value{
a coloc dataset
}
\description{
Create the standard dataset for the [coloc] package
}
