% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise.R
\name{standardise_data}
\alias{standardise_data}
\title{Standardise data inputs for processing}
\usage{
standardise_data(dat, source, build_from = NULL, build_to = NULL)
}
\arguments{
\item{dat}{a data.frame like object}

\item{source}{a string, one of c("ebi_eqtl_api")}

\item{build_from}{valid parameter for `genepi.utils::lift()`}

\item{build_to}{valid parameter for `genepi.utils::lift()`}
}
\value{
a data.table
}
\description{
Minimum dataset for GWAS: c(RSID, CHR, BP, EA, OA, EAF, BETA, SE, P, N)
}
