% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TargetExplorer-package.R
\docType{package}
\name{TargetExplorer-package}
\alias{TargetExplorer}
\alias{TargetExplorer-package}
\title{TargetExplorer: Exploring Genetically Proxied Drug Targets}
\description{
An Rshiny interface to explore genetically proxied drug targets, with multiple lines of evidence.
}
\author{
\strong{Maintainer}: Nicholas Sunderland \email{nicholas.sunderland@bristol.ac.uk}

}
\keyword{internal}
