% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebi_eqtl_catalogue.R
\name{import_ebi_eqtl}
\alias{import_ebi_eqtl}
\title{Import EBI eQTL data}
\usage{
import_ebi_eqtl(
  study_info_table,
  studies = NULL,
  tissues = NULL,
  chr = NULL,
  bp_start = NULL,
  bp_end = NULL,
  cptid = NULL,
  rsid = NULL,
  gene_id = NULL,
  nlog10p = NULL,
  build = "GRCh37"
)
}
\arguments{
\item{study_info_table}{.}

\item{studies}{.}

\item{tissues}{.}

\item{chr}{.}

\item{bp_start}{.}

\item{bp_end}{.}

\item{cptid}{.}

\item{rsid}{.}

\item{gene_id}{.}

\item{nlog10p}{.}

\item{build}{.}
}
\value{
a data.frame
}
\description{
Import EBI eQTL data
}
